+++ packages/unstyled/menu/src/MenuItem.tsx
@@ -85,6 +85,18 @@ export function MenuItem({
         selected: state.selectionManager.isSelected(item.key),
         disabled: state.selectionManager.isDisabled(item.key),
       }}
+      dataSet={{
+        hover: isHovered ? 'true' : 'false',
+        focus: isFocused ? 'true' : 'false',
+        active: isPressed ? 'true' : 'false',
+        focusVisible: isFocusVisible ? 'true' : 'false',
+        selected: state.selectionManager.isSelected(item.key)
+          ? 'true'
+          : 'false',
+        disabled: state.selectionManager.isDisabled(item.key)
+          ? 'true'
+          : 'false',
+      }}
       {...rest}
       // @ts-ignore - web only
       onHoverIn={composeEventHandlers(rest?.onHoverIn, hoverProps.onHoverIn)}