+++ packages/unstyled/slider/src/SliderTrack.tsx
@@ -39,6 +39,13 @@ function SliderTrack<StyledSliderTrackProps>(
           focusVisible: isFocusVisible || isFocusVisibleProp,
           active: isPressed || isPressedProp,
         }}
+        dataSet={{
+          hover: isHovered || isHoveredProp ? 'true' : 'false',
+          disabled: isDisabled || isDisabledProp ? 'true' : 'false',
+          focus: isFocused || isFocusedProp ? 'true' : 'false',
+          focusVisible: isFocusVisible || isFocusVisibleProp ? 'true' : 'false',
+          active: isPressed || isPressedProp ? 'true' : 'false',
+        }}
         disabled={isDisabled}
       >
         {children}