+++ packages/unstyled/accordion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gluestack-ui/accordion
 
+## 1.0.1
+
+### Patch Changes
+
+- Feature
+  Added Support of data-state-\* props
+
 ## 1.0.0
 
 ### Major Changes