+++ packages/unstyled/radio/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gluestack-ui/radio
 
+## 0.1.25
+
+### Patch Changes
+
+- Feature
+  Added Support of data-state-\* props
+- Updated dependencies
+  - @gluestack-ui/form-control@0.1.16
+
 ## 0.1.24
 
 ### Patch Changes